+++ bin/docker/images/woocommerce/setup-woocommerce.sh
@@ -43,10 +43,6 @@ if [ ! -f /var/www/html/wp-content/plugins/woocommerce/woocommerce.php ]; then
     printf "\n${COLOR_SUCCESS} ======================================= ${NC}\n"
     cd /var/www/html/wp-content/plugins/woocommerce_hipayenterprise/
 
-    cp composer.json composer.json.bak
-    cat composer.json.bak | python -c "import sys, json; composerObj=json.load(sys.stdin); composerObj['scripts'] = {'post-install-cmd': ['@managePiDataURLDev'], 'post-update-cmd': ['@managePiDataURLDev'], 'managePiDataURLDev': [\"sed -i 's@https://stage-data.hipay.com@"$PI_DATA_URL"@g' vendor/hipay/hipay-fullservice-sdk-php/lib/HiPay/Fullservice/HTTP/Configuration/Configuration.php\", \"sed -i 's@https://data.hipay.com@"$PI_DATA_URL"@g' vendor/hipay/hipay-fullservice-sdk-php/lib/HiPay/Fullservice/HTTP/Configuration/Configuration.php\"]}; print json.dumps(composerObj, False, True, True, True, None, 2);" >composer.json
-    rm composer.json.bak
-
     composer install --no-dev
 
     printf "\n${COLOR_SUCCESS} ======================================= ${NC}\n"
@@ -145,13 +141,6 @@ printf "${COLOR_SUCCESS}    |   URL BACK        : $WORDPRESS_URL/wp-admin       
 printf "${COLOR_SUCCESS}    |                                                                      ${NC}\n"
 printf "${COLOR_SUCCESS}    |   PHP VERSION     : $PHP_VERSION                                     ${NC}\n"
 printf "${COLOR_SUCCESS}    |======================================================================${NC}\n"
-
-printf "\n${COLOR_SUCCESS} ======================================= ${NC}\n"
-printf "\n${COLOR_SUCCESS}           HOSTS CONGIGURATION           ${NC}\n"
-printf "\n${COLOR_SUCCESS} ======================================= ${NC}\n"
-cp /etc/hosts ~/hosts.bak
-sed -i 's/^127\.0\.0\.1.*/127.0.0.1    localhost    data.hipay.com    stage-data.hipay.com/g' ~/hosts.bak
-cp ~/hosts.bak /etc/hosts
 
 #==========================================
 # APACHE RUNNING